requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
black==22.10.0
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
Django==4.1.4
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
frozenlist==1.3.3
idna==3.4
multidict==6.0.3
mypy-extensions==0.4.3
pathspec==0.10.2
platformdirs==2.6.0
pytz==2022.6
sqlparse==0.4.3
tomli==2.0.1
tzdata==2022.7
yarl==1.8.2
